# server/spec/openapi/openapi.yaml
...
openapi: "3.0.3"
info:
  title: BenxinAdminPro API
  description: |
    本心通用管理后台底座 API 规范。
    v0.1.0: 安全头、C 端加密信封格式和统一安全错误码段。
    v0.2.0: 认证接口（登录/刷新/登出/验证码）。
    v0.3.0: 组织架构（用户/部门/岗位 CRUD）。
    v0.4.0: RBAC 核心（角色/菜单/权限 + Casbin 联动 + Hashid）。
  version: "0.4.0"
  contact:
    name: 仗键天涯(daxing)
    email: 3442535897@qq.com
servers:
  - url: http://localhost:8080
    description: 本地开发

# ---------------------------------------------------------------------------
# 安全定义
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 0=成功，其它=错误码
        message:
          type: string
        data:
          description: 业务数据
...
    SysUser:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        nickname: { type: string }
        avatar: { type: string }
        email: { type: string }
        mobile: { type: string }
        dept_id: { type: integer, format: int64, nullable: true }
        status: { type: integer }
        remark: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PageResult:
      type: object
      properties:
        list: { type: array, items: {} }
        total: { type: integer, format: int64 }
        page: { type: integer }
        page_size: { type: integer }
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
        nickname: { type: string }
        email: { type: string }
        mobile: { type: string }
        dept_id: { type: integer, format: int64 }
        status: { type: integer }
        post_ids:
          type: array
          items: { type: integer, format: int64 }
...
paths:
...
  /sys/users:
    get:
      tags: [System]
      summary: 用户分页列表
      operationId: listUsers
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: username, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: integer } }
        - { name: dept_id, in: query, schema: { type: integer } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: 用户列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/PageResult" }
    post:
      tags: [System]
      summary: 新建用户
      operationId: createUser
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "200":
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SysUser" }
  /sys/users/{id}:
    get:
      tags: [System]
      summary: 用户详情
      operationId: getUser
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: 用户详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SysUser" }
    put:
      tags: [System]
      summary: 更新用户
      operationId: updateUser
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SysUser" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
    delete:
      tags: [System]
      summary: 删除用户
      operationId: deleteUser
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /sys/users/{id}/password:
    put:
      tags: [System]
      summary: 重置密码
      operationId: resetUserPassword
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "200":
          description: 重置成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /sys/users/{id}/status:
    put:
      tags: [System]
      summary: 启用/禁用用户
      operationId: setUserStatus
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: integer }
      responses:
        "200":
          description: 设置成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
...
  /sys/users/{id}/roles:
    put:
      tags: [RBAC]
      summary: 分配用户角色
      operationId: assignUserRoles
      security: [{ BearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role_ids: { type: array, items: { type: integer } }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
...